♻️ Refactor: VoiceRangeDisplay 파일 분리 #52
